chore: rename bdk crate to bdk_wallet
