+++ crates/hwi/Cargo.toml
@@ -9,5 +9,5 @@ license = "MIT OR Apache-2.0"
 readme = "README.md"
 
 [dependencies]
-bdk = { path = "../bdk" }
+bdk_wallet = { path = "../wallet" }
 hwi = { version = "0.8.0", features = [ "miniscript"] }
