+++ crates/wallet/src/descriptor/error.rs
