+++ crates/wallet/src/lib.rs
