+++ crates/wallet/src/wallet/error.rs
